requirements.txt
discord.py
yt-dlp
PyNaCl
python-dotenv
ffmpeg-python

brotli
static-ffmpeg
aiosqlite
